+++ .idea/scala_compiler.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ScalaCompilerConfiguration">
-    <profile name="SBT 1" modules="scalashell,scala-shell" />
+    <profile name="SBT 1" modules="scala-shell" />
   </component>
 </project>
