+++ src/main/scala/exceptions/InterpretationExceptions.scala
@@ -1,6 +1,12 @@
 package exceptions
 
 
+/**
+  * Error that occurred during execution of the external command.
+  *
+  * If execution of that command ended by throwing exception, then
+  * `cause` will contain that exception.
+  */
 case class ExternalCommandErrorException(
     message: String = null,
     cause: Throwable = null,
