+++ src/main/scala/interpretation/Interpreter.scala
@@ -1,7 +1,9 @@
 package interpretation
 
 import environment.Context
+import interpretation.commands.CommandsFactory
 import io._
+import io.streams._
 import language.AST._
 
 import scala.collection.mutable.ListBuffer
