+++ src/main/scala/io/streams/FiniteStream.scala
@@ -1,0 +1,6 @@
+package io.streams
+
+trait FiniteStream[T] extends InputStream[T] {
+  def toList: List[T]
+  def isAtEnd(): Boolean
+}
