+++ src/main/scala/io/streams/InputStream.scala
@@ -1,0 +1,6 @@
+package io.streams
+
+trait InputStream[T] {
+  def read(): Option[T]
+  def toJavaInputStream: java.io.InputStream
+}
