+++ src/main/scala/io/streams/OutputStream.scala
@@ -1,0 +1,6 @@
+package io.streams
+
+trait OutputStream[T] {
+  def write(value: T)
+  def toJavaOutputStream : java.io.OutputStream
+}
