--- src/main/scala/language/AST/Visitor.scala
@@ -1,7 +1,0 @@
-package language.AST
-
-abstract class Visitor {
-  def visit(node: Node): Any = { }
-  def previsit(node: Node): Any = { }
-  def postvisit(node: Node): Any = { }
-}
